move user API functions
