+++ py_jama_client/client.py
@@ -47,7 +47,7 @@ class BaseClient:
     def get_all(
         self,
         resource,
-        params=None,
+        params: Optional[dict] = None,
         allowed_results_per_page=DEFAULT_ALLOWED_RESULTS_PER_PAGE,
         **kwargs,
     ):
@@ -1310,78 +1310,6 @@ class JamaClient(BaseClient):
             **kwargs,
         )
 
-    def get_users(
-        self,
-        *args,
-        params: Optional[dict] = None,
-        allowed_results_per_page=DEFAULT_ALLOWED_RESULTS_PER_PAGE,
-        **kwargs,
-    ):
-        """
-        Gets a list of all active users visible to the current user
-
-        Args:
-            allowed_results_per_page: Number of results per page
-
-        Returns: JSON array
-
-        """
-        resource_path = "users/"
-        return self.get_all(
-            resource_path,
-            params,
-            allowed_results_per_page=allowed_results_per_page,
-            **kwargs,
-        )
-
-    def get_user(
-        self,
-        user_id: int,
-        *args,
-        params: Optional[dict] = None,
-        **kwargs,
-    ):
-        """
-        Gets a single speificed user
-
-        Args:
-            user_id: user api ID
-
-        Returns: JSON obect
-
-        """
-        resource_path = f"users/{user_id}"
-        try:
-            response = self._core.get(
-                resource_path,
-                params,
-                **kwargs,
-            )
-        except CoreException as err:
-            py_jama_rest_client_logger.error(err)
-            raise APIException(str(err))
-        return ClientResponse.from_response(response)
-
-    def get_current_user(
-        self,
-        *args,
-        params: Optional[dict] = None,
-        **kwargs,
-    ):
-        """
-        Gets a current user
-
-        Returns: JSON obect
-
-        """
-        resource_path = "users/current"
-        try:
-            response = self._core.get(resource_path, params, **kwargs)
-        except CoreException as err:
-            py_jama_rest_client_logger.error(err)
-            raise APIException(str(err))
-        return ClientResponse.from_response(response)
-
     def get_test_cycle(
         self,
         test_cycle_id: int,
@@ -1490,66 +1418,6 @@ class JamaClient(BaseClient):
 
         BaseClient.handle_response_status(response)
         return response.status_code
-
-    def post_user(
-        self,
-        username: str,
-        password: str,
-        first_name: str,
-        last_name: str,
-        email: str,
-        license_type: str,
-        phone: str = None,
-        title: str = None,
-        location: str = None,
-        *args,
-        params: Optional[dict] = None,
-        **kwargs,
-    ):
-        """
-        Creates a new user
-
-        Args:
-            username: str
-            password: str
-            first_name: str
-            last_name: str
-            email: str
-            phone: str - optional
-            title: str - optional
-            location: str - optional
-            licenseType: enum [ NAMED, FLOATING, STAKEHOLDER, FLOATING_COLLABORATOR, RESERVED_COLLABORATOR, FLOATING_REVIEWER, RESERVED_REVIEWER, NAMED_REVIEWER, TEST_RUNNER, EXPIRING_TRIAL, INACTIVE ]
-
-        Returns: newly created user
-
-        """
-
-        body = {
-            "username": username,
-            "password": password,
-            "firstName": first_name,
-            "lastName": last_name,
-            "email": email,
-            "phone": phone,
-            "title": title,
-            "location": location,
-            "licenseType": license_type,
-        }
-        resource_path = "users/"
-        headers = {"content-type": "application/json"}
-        try:
-            response = self._core.post(
-                resource_path,
-                params,
-                data=json.dumps(body),
-                headers=headers,
-                **kwargs,
-            )
-        except CoreException as err:
-            py_jama_rest_client_logger.error(err)
-            raise APIException(str(err))
-        BaseClient.handle_response_status(response)
-        return ClientResponse.from_response(response)
 
     def post_tag(
         self,
@@ -2032,98 +1900,6 @@ class JamaClient(BaseClient):
             except CoreException as err:
                 py_jama_rest_client_logger.error(err)
                 raise APIException(str(err))
-        self.handle_response_status(response)
-        return response.status_code
-
-    def put_user(
-        self,
-        user_id: int,
-        username: str,
-        password: str,
-        first_name: str,
-        last_name: str,
-        email: str,
-        phone: str = None,
-        title: str = None,
-        location: str = None,
-        *args,
-        params: Optional[dict] = None,
-        **kwargs,
-    ):
-        """
-        updates an existing user
-
-        Args:
-            username: str
-            password: str
-            first_name: str
-            last_name: str
-            email: str
-            phone: str - optional
-            title: str - optional
-            location: str - optional
-
-        Returns: api status code
-
-        """
-
-        body = {
-            "username": username,
-            "password": password,
-            "firstName": first_name,
-            "lastName": last_name,
-            "email": email,
-            "phone": phone,
-            "title": title,
-            "location": location,
-        }
-        resource_path = f"users/{user_id}"
-        headers = {"content-type": "application/json"}
-        try:
-            response = self._core.put(
-                resource_path,
-                params,
-                data=json.dumps(body),
-                headers=headers,
-                **kwargs,
-            )
-        except CoreException as err:
-            py_jama_rest_client_logger.error(err)
-            raise APIException(str(err))
-        self.handle_response_status(response)
-        return response.status_code
-
-    def put_user_active(
-        self,
-        user_id: int,
-        is_active: bool,
-        *args,
-        params: Optional[dict] = None,
-        **kwargs,
-    ):
-        """
-        updates an existing users active status
-
-        Args:
-            is_active: boolean
-
-        Returns: api status code
-
-        """
-        body = {"active": is_active}
-        resource_path = f"users/{user_id}/active"
-        headers = {"content-type": "application/json"}
-        try:
-            response = self._core.put(
-                resource_path,
-                params,
-                data=json.dumps(body),
-                headers=headers,
-                **kwargs,
-            )
-        except CoreException as err:
-            py_jama_rest_client_logger.error(err)
-            raise APIException(str(err))
         self.handle_response_status(response)
         return response.status_code
